Changed the way in which Reader parallelism works in DataGen. Now the number of parallel threads in reader is configurable, and each thread does a % on the file name and reads the file as per its readEach number assigned. This keeps the threads busy till end of execution.
In Sankar machine it took 35 mins to complete DataGen. That is appx. 10 mins faster than before.

Note: This code expects all the Object files to reside directly in OBJECTS folder. Consult Sankar how to make this happen.
I ran it with 11 reader threads and 16 reader threads. No big performance difference. Maybe there could be some difference depending on the number of cores in the machine.